Done CSS and restyling

+++ src/proveedorControllers/ProductoProveedorViewController.java
@@ -1,17 +1,22 @@
 package proveedorControllers;
 
 import java.net.URL;
+import java.sql.SQLException;
 import java.util.ResourceBundle;
 import java.util.stream.Collectors;
 
 import application.Main;
+import bbdd_tools.ProveedorManager;
 import javafx.collections.FXCollections;
 import javafx.collections.ObservableList;
 import javafx.fxml.FXML;
 import javafx.fxml.Initializable;
+import javafx.scene.control.Alert;
+import javafx.scene.control.ButtonType;
 import javafx.scene.control.SelectionMode;
 import javafx.scene.control.TableColumn;
 import javafx.scene.control.TableView;
+import javafx.scene.control.Alert.AlertType;
 import javafx.scene.control.cell.PropertyValueFactory;
 import performers.Producto;
 import tools.UITools;
@@ -40,8 +45,7 @@ public class ProductoProveedorViewController implements Initializable {
 		TableColumn<Producto, Float> descuentoProducto = new TableColumn<Producto, Float>("descuento");
 		descuentoProducto.setCellValueFactory(new PropertyValueFactory<Producto, Float>("descuento"));
 
-		productosData = FXCollections.observableList(Main.productos.stream()
-				.filter(e -> e.getProveedor().equals(Main.selectedProveedor)).collect(Collectors.toList()));
+		productosData = FXCollections.observableList(Main.selectedProveedor.getProductos());
 		productosTableView.getColumns().addAll(nombreProducto, ventaPorPesoProducto, precioProducto, stockProducto,
 				descuentoProducto);
 
@@ -53,7 +57,22 @@ public class ProductoProveedorViewController implements Initializable {
 	}
 	
 	public void deleteItems() {
+		ObservableList<Producto> itemsToDelete = productosTableView.getSelectionModel().getSelectedItems();
+		if(productosTableView.getSelectionModel().getSelectedItem() == null) {
+			new Alert(AlertType.ERROR, "¡Debes seleccionar un elemento primero!", ButtonType.OK).show();
+			return;
+		}
+		
+		
+		try {
+			ProveedorManager.deleteProduct(productosTableView.getSelectionModel().getSelectedItems());
+		} catch (SQLException e) {
+			e.printStackTrace();
+		}
+		
 		UITools.deleteItemsFromControl(productosTableView);
+		Main.selectedProveedor.getProductos().removeAll(itemsToDelete);
+		
 	}
 
 }
